Move vibronic into new qchem module (#416)
* Move vibronic module

* Remove whitespace

* Remove vibronic from apps init

* Update in vibronic tests

* Move vibronic tests

* Remove vibronic from sample module

* Remove vibronic test from sample tests

* Add vibronic sampler to qchem.vibronic

* finish moving in tests

* Rename vibronic() to sample()

* Update tutorial

* Update changelog

* Tidy up mentions of vibronic module

* Update tutorial

* Fixing docs

* Update changelog with PR

* Apply black

* Add imports for backward compatibility

* Add pylint ignore

* Update changelog

* Shorten function reference
