+++ strawberryfields/apps/__init__.py
@@ -27,7 +27,6 @@ Strawberry Fields application layer.
 	similarity
 	subgraph
 	train
-	vibronic
 """
 
 import strawberryfields.apps.clique
@@ -39,4 +38,4 @@ import strawberryfields.apps.sample
 import strawberryfields.apps.similarity
 import strawberryfields.apps.subgraph
 import strawberryfields.apps.train
-import strawberryfields.apps.vibronic
+from strawberryfields.apps.qchem import vibronic
