+++ strawberryfields/apps/sample.py
@@ -41,9 +41,6 @@ noise in quantum photonics. Here, ``loss = 0`` describes ideal loss-free GBS, wh
 Samples can be postselected based upon their total number of photons or clicks and the ``numpy``
 random seed used to generate samples can be fixed.
 
-The :func:`vibronic` function allows users to sample Gaussian states for computing molecular
-vibronic spectra.
-
 Generating subgraphs
 --------------------
 
@@ -110,6 +107,7 @@ import networkx as nx
 import numpy as np
 
 import strawberryfields as sf
+from strawberryfields.apps.qchem.vibronic import sample as vibronic  # pylint: disable=unused-import
 
 
 def sample(
@@ -274,103 +272,6 @@ def to_subgraphs(samples: list, graph: nx.Graph) -> list:
         return [sorted([graph_nodes[i] for i in s]) for s in subgraph_samples]
 
     return subgraph_samples
-
-
-def vibronic(
-    t: np.ndarray,
-    U1: np.ndarray,
-    r: np.ndarray,
-    U2: np.ndarray,
-    alpha: np.ndarray,
-    n_samples: int,
-    loss: float = 0.0,
-) -> list:
-    """Generate samples for computing vibronic spectra. The following gates are applied to input
-    vacuum states:
-
-    1. Two-mode squeezing on all :math:`2N` modes with parameters ``t``
-    2. Interferometer ``U1`` on the first :math:`N` modes
-    3. Squeezing on the first :math:`N` modes with parameters ``r``
-    4. Interferometer ``U2`` on the first :math:`N` modes
-    5. Displacement on the first :math:`N` modes with parameters ``alpha``
-
-    A sample is generated by measuring the number of photons in each of the :math:`2N` modes. In
-    the special case that all of the two-mode squeezing parameters ``t`` are zero, only :math:`N`
-    modes are considered, which speeds up calculations.
-
-    **Example usage:**
-
-    >>> formic = data.Formic()
-    >>> w = formic.w
-    >>> wp = formic.wp
-    >>> Ud = formic.Ud
-    >>> delta = formic.delta
-    >>> T = 0
-    >>> t, U1, r, U2, alpha = vibronic.gbs_params(w, wp, Ud, delta, T)
-    >>> sample.vibronic(t, U1, r, U2, alpha, 2, 0.0)
-    [[0, 0, 2, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0],
-     [0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]]
-
-    Args:
-        t (array): two-mode squeezing parameters
-        U1 (array): unitary matrix for the first interferometer
-        r (array): squeezing parameters
-        U2 (array): unitary matrix for the second interferometer
-        alpha (array): displacement parameters
-        n_samples (int): number of samples to be generated
-        loss (float): loss parameter denoting the fraction of generated photons that are lost
-
-
-    Returns:
-        list[list[int]]: a list of samples from GBS
-    """
-    if n_samples < 1:
-        raise ValueError("Number of samples must be at least one")
-    if not 0 <= loss <= 1:
-        raise ValueError("Loss parameter must take a value between zero and one")
-
-    n_modes = len(t)
-
-    eng = sf.LocalEngine(backend="gaussian")
-
-    if np.any(t != 0):
-        gbs = sf.Program(n_modes * 2)
-    else:
-        gbs = sf.Program(n_modes)
-
-    # pylint: disable=expression-not-assigned,pointless-statement
-    with gbs.context as q:
-
-        if np.any(t != 0):
-            for i in range(n_modes):
-                sf.ops.S2gate(t[i]) | (q[i], q[i + n_modes])
-
-        sf.ops.Interferometer(U1) | q[:n_modes]
-
-        for i in range(n_modes):
-            sf.ops.Sgate(r[i]) | q[i]
-
-        sf.ops.Interferometer(U2) | q[:n_modes]
-
-        for i in range(n_modes):
-            sf.ops.Dgate(np.abs(alpha[i]), np.angle(alpha[i])) | q[i]
-
-        if loss:
-            for _q in q:
-                sf.ops.LossChannel(1 - loss) | _q
-
-        sf.ops.MeasureFock() | q
-
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", category=UserWarning, message="Cannot simulate non-")
-
-        s = eng.run(gbs, shots=n_samples).samples
-
-    s = np.array(s).tolist()  # convert all generated samples to list
-
-    if np.any(t == 0):
-        s = np.pad(s, ((0, 0), (0, n_modes))).tolist()
-    return s
 
 
 def waw_matrix(A: np.ndarray, w: Union[np.ndarray, list]) -> np.ndarray:
